GGally-package {GGally}	R Documentation

GGally: Extension to 'ggplot2'

Description

The R package 'ggplot2' is a plotting system based on the grammar of graphics. 'GGally' extends 'ggplot2' by adding several functions to reduce the complexity of combining geometric objects with transformed data. Some of these functions include a pairwise plot matrix, a two group pairwise plot matrix, a parallel coordinates plot, a survival plot, and several functions to plot networks.

Author(s)

Maintainer: Barret Schloerke schloerke@gmail.com

Authors:

Di Cook dicook@monash.edu [thesis advisor]

Joseph Larmarange joseph@larmarange.net

Francois Briatte f.briatte@gmail.com

Moritz Marbach mmarbach@mail.uni-mannheim.de

Edwin Thoen edwinthoen@gmail.com

Amos Elberg amos.elberg@gmail.com

Jason Crowley crowley.jason.s@gmail.com

Other contributors:

Ott Toomet otoomet@gmail.com [contributor]

Heike Hofmann hofmann@iastate.edu [thesis advisor]

Hadley Wickham h.wickham@gmail.com [thesis advisor]

See Also

Useful links:

https://ggobi.github.io/ggally/

https://github.com/ggobi/ggally

Report bugs at https://github.com/ggobi/ggally/issues

[Package GGally version 2.1.2 Index]