Programme for International Student Assessment (PISA) 2012 Data for Australia

About PISA

Usage

data(australia_PISA2012)

Format

A data frame with 8247 rows and 32 variables

Source

https://www.oecd.org/en/data/datasets/pisa-2012-cba-database.html

Details

The Programme for International Student Assessment (PISA) is a triennial international survey which aims to evaluate education systems worldwide by testing the skills and knowledge of 15-year-old students. To date, students representing more than 70 economies have participated in the assessment.

While 65 economies took part in the 2012 study, this data set only contains information from the country of Australia.

gender : Factor w/ 2 levels "female","male": 1 1 2 2 2 1 1 1 2 1 ...

age : Factor w/ 4 levels "4","5","6","7": 2 2 2 4 3 1 2 2 2 2 ...

homework : num 5 5 9 3 2 3 4 3 5 1 ...

desk : num 1 0 1 1 1 1 1 1 1 1 ...

room : num 1 1 1 1 1 1 1 1 1 1 ...

study : num 1 1 1 1 1 1 1 1 1 1 ...

computer : num 1 1 1 1 1 1 1 1 1 1 ...

software : num 1 1 1 1 1 1 1 1 1 1 ...

internet : num 1 1 1 1 1 1 1 1 1 1 ...

literature : num 0 0 1 0 1 1 1 1 1 0 ...

poetry : num 0 0 1 0 1 1 0 1 1 1 ...

art : num 1 0 1 0 1 1 0 1 1 1 ...

textbook : num 1 1 1 1 1 0 1 1 1 1 ...

dictionary : num 1 1 1 1 1 1 1 1 1 1 ...

dishwasher : num 1 1 1 1 0 1 1 1 1 1 ...

PV1MATH : num 562 565 602 520 613 ...

PV2MATH : num 569 557 594 507 567 ...

PV3MATH : num 555 553 552 501 585 ...

PV4MATH : num 579 538 526 521 596 ...

PV5MATH : num 548 573 619 547 603 ...

PV1READ : num 582 617 650 554 605 ...

PV2READ : num 571 572 608 560 557 ...

PV3READ : num 602 560 594 517 627 ...

PV4READ : num 572 564 575 564 597 ...

PV5READ : num 585 565 620 572 598 ...

PV1SCIE : num 583 627 668 574 639 ...

PV2SCIE : num 579 600 665 612 635 ...

PV3SCIE : num 593 574 620 571 666 ...

PV4SCIE : num 567 582 592 598 700 ...

PV5SCIE : num 587 625 656 662 670 ...

SENWGT_STU : num 0.133 0.133 0.141 0.141 0.141 ...

possessions: num 10 8 12 9 11 11 10 12 12 11 ...