flea {GGally}	R Documentation

Historical data used for classification examples.

Description

This data contains physical measurements on three species of flea beetles.

Usage

data(flea)

Format

A data frame with 74 rows and 7 variables

Details

species Ch. concinna, Ch. heptapotamica, Ch. heikertingeri

tars1 width of the first joint of the first tarsus in microns

tars2 width of the second joint of the first tarsus in microns

head the maximal width of the head between the external edges of the eyes in 0.01 mm

aede1 the maximal width of the aedeagus in the fore-part in microns

aede2 the front angle of the aedeagus (1 unit = 7.5 degrees)

aede3 the aedeagus width from the side in microns

References

Lubischew, A. A. (1962), On the Use of Discriminant Functions in Taxonomy, Biometrics 18:455-477.

[Package GGally version 2.1.2 Index]