ggally_facetdensitystrip {GGally}	R Documentation

Density or tiles plot with facets

Description

Make tile plot or density plot as compact as possible.

Usage

ggally_facetdensitystrip(data, mapping, ..., den_strip = FALSE)

Arguments

data
data set using

mapping
aesthetics being used

...
other arguments being sent to either geom_histogram or stat_density

den_strip
boolean to decide whether or not to plot a density strip(TRUE) or a facet density(FALSE) plot.

Author(s)

Barret Schloerke

Examples

example(ggally_facetdensity)
example(ggally_denstrip)

[Package GGally version 2.1.2 Index]