ggts {GGally}	R Documentation

Multiple time series

Description

GGally implementation of ts.plot. Wraps around the ggduo function and removes the column strips

Usage

ggts(..., columnLabelsX = NULL, xlab = "time")

Arguments

...
supplied directly to ggduo

columnLabelsX
remove top strips for the X axis by default

xlab
defaults to "time"

Value

ggmatrix object

Examples

# Small function to display plots only if it's interactive
p_ <- GGally::print_if_interactive

p_(ggts(pigs, "time", c("gilts", "profit", "s_per_herdsz", "production", "herdsz")))

[Package GGally version 2.1.2 Index]