is_blank_plot {GGally}	R Documentation

Is Blank Plot? Find out if the plot equals a blank plot

Description

Is Blank Plot? Find out if the plot equals a blank plot

Usage

is_blank_plot(p)

Examples

 GGally:::is_blank_plot(ggally_blank())
 GGally:::is_blank_plot(ggally_points(mtcars, ggplot2::aes_string(x = "disp", y = "hp")))

[Package GGally version 2.1.2 Index]