lowertriangle {GGally}	R Documentation

lowertriangle - rearrange dataset as the preparation of ggscatmat function

Description

function for making the melted dataset used to plot the lowertriangle scatterplots.

Usage

lowertriangle(data, columns = 1:ncol(data), color = NULL)

Arguments

data
a data matrix. Should contain numerical (continuous) data.

columns
an option to choose the column to be used in the raw dataset. Defaults to 1:ncol(data)

color
an option to choose a factor variable to be grouped with. Defaults to (NULL)

Author(s)

Mengjia Ni, Di Cook

Examples

data(flea)
head(lowertriangle(flea, columns= 2:4))
head(lowertriangle(flea))
head(lowertriangle(flea, color="species"))

[Package GGally version 2.1.2 Index]