remove_color_unless_equal {GGally}	R Documentation

Remove colour mapping unless found in select mapping keys

Description

Remove colour mapping unless found in select mapping keys

Usage

remove_color_unless_equal(mapping, to = c("x", "y"))

Arguments

mapping
output of ggplot2::aes(...)

to
set of mapping keys to check

Value

Aes mapping with colour mapping kept only if found in selected mapping keys.

Examples

mapping <- aes(x = sex, y = age, colour = sex)

mapping <- aes(x = sex, y = age, colour = region)
remove_color_unless_equal(mapping)

[Package GGally version 2.1.2 Index]