scag_order {GGally}	R Documentation

Find order of variables

Description

Find order of variables based on a specified scagnostic measure by maximizing the index values of that measure along the path.

Usage

scag_order(scag, vars, measure)

Arguments

scag
scagnostics object

vars
character vector of the variables to be ordered

measure
scagnostics measure to order according to

Value

character vector of variable ordered according to the given scagnostic measure

Author(s)

Barret Schloerke

[Package GGally version 2.1.2 Index]