scatmat {GGally}	R Documentation

scatmat - plot the lowertriangle plots and density plots of the scatter plot matrix.

Description

function for making scatterplots in the lower triangle and diagonal density plots.

Usage

scatmat(data, columns = 1:ncol(data), color = NULL, alpha = 1)

Arguments

data
a data matrix. Should contain numerical (continuous) data.

columns
an option to choose the column to be used in the raw dataset. Defaults to 1:ncol(data)

color
an option to group the dataset by the factor variable and color them by different colors. Defaults to NULL

alpha
an option to set the transparency in scatterplots for large data. Defaults to 1.

Author(s)

Mengjia Ni, Di Cook dicook@monash.edu

Examples

data(flea)
scatmat(flea, columns=2:4)
scatmat(flea, columns= 2:4, color="species")

[Package GGally version 1.5.0 Index]