stat_cross {GGally}	R Documentation

Compute cross-tabulation statistics

Description

Computes statistics of a 2-dimensional matrix using augment.htest from broom.

Usage

stat_cross(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  keep.zero.cells = FALSE
)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

geom
Override the default connection between geom_point and stat_prop.

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

na.rm
If TRUE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

keep.zero.cells
If TRUE, cells with no observations are kept.

Aesthetics

stat_prop requires the x and the y aesthetics.

Computed variables

observed

number of observations in x,y

prop

proportion of total

row.prop

row proportion

col.prop

column proportion

expected

expected count under the null hypothesis

resid

Pearson's residual

std.resid

standardized residual

Examples

# Small function to display plots only if it's interactive
p_ <- GGally::print_if_interactive

d <- as.data.frame(Titanic)

# plot number of observations
p_(ggplot(d) +
 aes(x = Class, y = Survived, weight = Freq, size = after_stat(observed)) +
 stat_cross() +
 scale_size_area(max_size = 20))

# custom shape and fill colour based on chi-squared residuals
p_(ggplot(d) +
 aes(
   x = Class, y = Survived, weight = Freq,
   size = after_stat(observed), fill = after_stat(std.resid)
 ) +
 stat_cross(shape = 22) +
 scale_fill_steps2(breaks = c(-3, -2, 2, 3), show.limits = TRUE) +
 scale_size_area(max_size = 20))

# plotting the number of observations as a table
p_(ggplot(d) +
 aes(
   x = Class, y = Survived, weight = Freq, label = after_stat(observed)
 ) +
 geom_text(stat = "cross"))

# Row proportions with standardized residuals
p_(ggplot(d) +
  aes(
    x = Class, y = Survived, weight = Freq,
    label = scales::percent(after_stat(row.prop)),
    size = NULL, fill = after_stat(std.resid)
  ) +
  stat_cross(shape = 22, size = 30) +
  geom_text(stat = "cross") +
  scale_fill_steps2(breaks = c(-3, -2, 2, 3), show.limits = TRUE) +
  facet_grid(Sex ~ .) +
  labs(fill = "Standardized residuals") +
  theme_minimal())

# can work with continuous or character variables
data(tips, package = "reshape")
p_(ggplot(tips) +
  aes(x = tip, y = as.character(day), size = after_stat(observed)) +
  stat_cross(alpha = .1, color = "blue") +
  scale_size_area(max_size = 12))

[Package GGally version 2.1.2 Index]