GGally: Extension to 'ggplot2'

The R package 'ggplot2' is a plotting system based on the grammar of graphics. 'GGally' extends 'ggplot2' by adding several functions to reduce the complexity of combining geometric objects with transformed data. Some of these functions include a pairwise plot matrix, a two group pairwise plot matrix, a parallel coordinates plot, a survival plot, and several functions to plot networks.

See also

Author

Maintainer: Barret Schloerke schloerke@gmail.com (ORCID)

Authors:

Di Cook dicook@monash.edu (ORCID) [thesis advisor]

Joseph Larmarange joseph@larmarange.net (ORCID)

Francois Briatte f.briatte@gmail.com

Moritz Marbach mmarbach@mail.uni-mannheim.de

Edwin Thoen edwinthoen@gmail.com

Amos Elberg amos.elberg@gmail.com

Jason Crowley crowley.jason.s@gmail.com

Other contributors:

Ott Toomet otoomet@gmail.com [contributor]

Heike Hofmann hhofmann4@unl.edu (ORCID) [thesis advisor]

Hadley Wickham h.wickham@gmail.com (ORCID) [thesis advisor]