uppertriangle {GGally}	R Documentation

Rearrange dataset as the preparation of ggscatmat function

Description

Function for making the dataset used to plot the uppertriangle plots.

Usage

uppertriangle(
  data,
  columns = 1:ncol(data),
  color = NULL,
  corMethod = "pearson"
)

Arguments

data
a data matrix. Should contain numerical (continuous) data.

columns
an option to choose the column to be used in the raw dataset. Defaults to 1:ncol(data)

color
an option to choose a factor variable to be grouped with. Defaults to (NULL)

corMethod
method argument supplied to cor

Author(s)

Mengjia Ni, Di Cook

Examples

data(flea)
head(uppertriangle(flea, columns=2:4))
head(uppertriangle(flea))
head(uppertriangle(flea, color="species"))

[Package GGally version 2.1.2 Index]