uppertriangle - rearrange dataset as the preparation of ggscatmat function

function for making the dataset used to plot the uppertriangle plots.

uppertriangle(
  data,
  columns = 1:ncol(data),
  color = NULL,
  corMethod = "pearson"
)

Arguments

data
a data matrix. Should contain numerical (continuous) data.

columns
an option to choose the column to be used in the raw dataset. Defaults to 1:ncol(data)

color
an option to choose a factor variable to be grouped with. Defaults to (NULL)

corMethod
method argument supplied to cor

Author

Mengjia Ni, Di Cook dicook@monash.edu

Examples

data(flea)
head(uppertriangle(flea, columns=2:4))

#>    xlab  ylab    r xvalue yvalue
#> 1  head tars1 -0.1   50.5  182.0
#> 2 tars2 tars1 0.03  126.5  182.0
#> 3  head tars2 0.67   50.5  126.5

head(uppertriangle(flea))

#>    xlab  ylab     r xvalue yvalue
#> 1 aede2 aede1 -0.25   12.0  136.5
#> 2 aede3 aede1  0.78   89.0  136.5
#> 3 aede3 aede2 -0.48   89.0   12.0
#> 4 aede1  head  0.59  136.5   50.5
#> 5 aede2  head -0.31   12.0   50.5
#> 6 aede3  head  0.52   89.0   50.5

head(uppertriangle(flea, color="species"))

#>    ylab  xlab colorcolumn     r xvalue yvalue
#> 1 aede1 aede2    Concinna -0.19     12 126.25
#> 2 aede1 aede2    Heikert. -0.06     12 136.50
#> 3 aede1 aede2   Heptapot. -0.12     12 146.75
#> 4 aede1 aede3    Concinna  0.45     89 126.25
#> 5 aede1 aede3    Heikert.  0.28     89 136.50
#> 6 aede1 aede3   Heptapot.  0.33     89 146.75